--- src/main/java/org/api/json/UserInfoResponse.java
@@ -1,23 +1,0 @@
-package org.api.json;
-
-public class UserInfoResponse {
-
-    private String email;
-    private String name;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
