--- src/main/java/org/api/json/UserRequest.java
@@ -1,39 +1,0 @@
-package org.api.json;
-
-public class UserRequest {
-    private String email;
-    private String password;
-    private String name;
-
-    public UserRequest(String email, String password, String name) {
-        this.email = email;
-        this.password = password;
-        this.name = name;
-    }
-
-    public UserRequest(){}
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
