--- src/main/java/org/api/json/UserResponse.java
@@ -1,40 +1,0 @@
-package org.api.json;
-
-public class UserResponse {
-
-    private Boolean success;
-    private String accessToken;
-    private String refreshToken;
-    private UserInfoResponse user;
-
-    public UserInfoResponse getUser() {
-        return user;
-    }
-
-    public void setUser(UserInfoResponse user) {
-        this.user = user;
-    }
-    public Boolean getSuccess() {
-        return success;
-    }
-
-    public void setSuccess(Boolean success) {
-        this.success = success;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-
-    public void setRefreshToken(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-}
