+++ src/main/java/site/nomoreparties/stellarburgers/json/UserInfoResponse.java
@@ -1,0 +1,15 @@
+package site.nomoreparties.stellarburgers.json;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class UserInfoResponse {
+
+    @Getter
+    @Setter
+    private String email;
+
+    @Getter
+    @Setter
+    private String name;
+}
