+++ src/main/java/site/nomoreparties/stellarburgers/json/UserResponse.java
@@ -1,0 +1,21 @@
+package site.nomoreparties.stellarburgers.json;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class UserResponse {
+
+    @Getter
+    @Setter
+    private Boolean success;
+    @Getter
+    @Setter
+    private String accessToken;
+    @Getter
+    @Setter
+    private String refreshToken;
+    @Getter
+    @Setter
+    private UserInfoResponse user;
+
+}
