+++ src/test/java/site/nomoreparties/stellarburgers/test/ConstructorTest.java
@@ -5,14 +5,9 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebDriver;
 import site.nomoreparties.stellarburgers.BrowserSetUp;
-import site.nomoreparties.stellarburgers.TestProperties;
 import site.nomoreparties.stellarburgers.pages.MainStellarBurgerPage;
-
-import java.util.Objects;
 
 public class ConstructorTest {
     BrowserSetUp browser = new BrowserSetUp();
